new fields in dto

+++ f1api/src/main/java/com/f1api/dto/LapTimeDTO.java
@@ -11,7 +11,7 @@ public class LapTimeDTO {
 
     private Short driverId;
 
-    private Short raceId;
+    private Long raceId;
 
     private Short lap;
